backend/requirements.txt
# Web Framework
fastapi>=0.103.1
uvicorn[standard]>=0.23.2
pydantic>=2.4.2
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Database
pymongo>=4.5.0
certifi>=2023.11.17

# Vector Search
qdrant-client>=1.6.0

# LLM Integration
langchain>=0.0.267
langchain-cohere>=0.1.0  # Added this missing dependency
langchain-google-genai>=0.1.1
google-generativeai>=0.3.1

# Utilities
aiohttp>=3.8.5
requests>=2.31.0
tenacity>=8.2.3
PyJWT>=2.8.0